+++ love/src/components/MainTel/MTDomePower/PowerPlot/PowerPlot.jsx
@@ -19,7 +19,6 @@ this program. If not, see <http://www.gnu.org/licenses/>.
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import PlotContainer from 'components/GeneralPurpose/Plot/Plot.container';
 import { VegaLite } from 'react-vega';
 import styles from './PowerPlot.module.css';
 
@@ -35,6 +34,8 @@ export default class PowerPlot extends Component {
     powerDrawLightWindScreen: PropTypes.number,
     /** Shutters power draw (W) */
     powerDrawShutter: PropTypes.number,
+    /** ClassName style to be asigned to the Power Plot */
+    className: PropTypes.string,
   };
 
   static defaultProps = {
@@ -155,41 +156,69 @@ export default class PowerPlot extends Component {
       // 4.0.0 DATA //
       ////////////////
       data: {
-        values: this.props.data,
-        // [
-        //   {
-        //     "system":"aCalibration Screen",
-        //     "count":powerDraw*0.35,
-        //     "date":"2020-01-01 10:10:00",
-        //   },{
-        //     "system":"aCalibration Screen",
-        //     "count":powerDraw*0.1,
-        //     "date":"2020-01-01 10:12:00",
-        //   },{
-        //     "system":"aCalibration Screen",
-        //     "count":powerDraw*0.05,
-        //     "date":"2020-01-01 10:14:00",
-        //   },{
-        //     "system":"aCalibration Screen",
-        //     "count":powerDraw*0.75,
-        //     "date":"2020-01-01 10:16:00",
-        //   },{
-        //     "system":"aCalibration Screen",
-        //     "count":powerDraw,
-        //     "date":"2020-01-01 10:18:00",
-        //   }
-        // ]
+        values: [
+          {
+            system: this.props.data[0]?.system ?? 0,
+            count: this.props.data[0]?.count ?? 0,
+            date: '2020-01-01 10:10:00',
+          },
+          {
+            system: this.props.data[1]?.system ?? 0,
+            count: this.props.data[1]?.count ?? 0,
+            date: '2020-01-01 10:11:00',
+          },
+          {
+            system: this.props.data[2]?.system ?? 0,
+            count: this.props.data[2]?.count ?? 0,
+            date: '2020-01-01 10:12:00',
+          },
+          {
+            system: this.props.data[3]?.system ?? 0,
+            count: this.props.data[3]?.count ?? 0,
+            date: '2020-01-01 10:13:00',
+          },
+          {
+            system: this.props.data[4]?.system ?? 0,
+            count: this.props.data[4]?.count ?? 0,
+            date: '2020-01-01 10:14:00',
+          },
+          {
+            system: this.props.data[5]?.system ?? 0,
+            count: this.props.data[5]?.count ?? 0,
+            date: '2020-01-01 10:15:00',
+          },
+          {
+            system: this.props.data[6]?.system ?? 0,
+            count: this.props.data[6]?.count ?? 0,
+            date: '2020-01-01 10:16:00',
+          },
+          {
+            system: this.props.data[7]?.system ?? 0,
+            count: this.props.data[7]?.count ?? 0,
+            date: '2020-01-01 10:17:00',
+          },
+          {
+            system: this.props.data[8]?.system ?? 0,
+            count: this.props.data[8]?.count ?? 0,
+            date: '2020-01-01 10:18:00',
+          },
+          {
+            system: this.props.data[9]?.system ?? 0,
+            count: this.props.data[9]?.count ?? 0,
+            date: '2020-01-01 10:19:00',
+          },
+        ],
       },
     };
 
     return (
-      <div className={styles.container}>
+      <div className={[styles.container, this.props.className].join(' ')}>
         <div className={styles.plot}>
           <VegaLite
             style={{ display: 'flex' }}
             renderer="svg"
             spec={spec}
-            className={[styles.plotContainer, this.props.className].join(' ')}
+            className={styles.plotContainer}
             actions={false}
           />
           <div style={{ width: 0 }}>